[front] multi-fix - start to create update dialog

See merge request la-fabrique-numerique/biocarburants!32

+++ web/api/v4/sanity_checks.py
@@ -63,14 +63,18 @@ def generic_error(error, **kwargs):
     d.update(kwargs)
     return GenericError(**d)
 
-def bulk_sanity_checks(lots, prefetched_data=None):
+def bulk_sanity_checks(lots, prefetched_data=None, dry_run=False):
     if not prefetched_data:
         prefetched_data = get_prefetched_data()
+    
     results = []
     errors = []
+    
     # cleanup previous errors
-    lot_ids = [l.id for l in lots]
-    GenericError.objects.filter(lot_id__in=lot_ids).delete()
+    if not dry_run:
+        lot_ids = [l.id for l in lots]
+        GenericError.objects.filter(lot_id__in=lot_ids).delete()
+    
     for lot in lots:
         try:
             is_sane, sanity_errors = sanity_check(lot, prefetched_data)
@@ -78,8 +82,11 @@ def bulk_sanity_checks(lots, prefetched_data=None):
             results.append(is_sane)
         except:
             traceback.print_exc()
-    GenericError.objects.bulk_create(errors, batch_size=1000)
-    return results
+    
+    if not dry_run:
+        GenericError.objects.bulk_create(errors, batch_size=1000)
+    
+    return errors, results
 
 
 def bulk_scoring(lots, prefetched_data=None):
@@ -198,7 +205,7 @@ def check_certificates(prefetched_data, lot, errors):
         else:
             # certificate is set and exists. is it valid?
             c = prefetched_data['certificates'][cert]
-            if c.valid_until < lot.delivery_date:
+            if c["valid_until"] < lot.delivery_date:
                  errors.append(generic_error(error=CarbureCertificatesErrors.EXPIRED_PRODSITE_CERT, lot=lot, display_to_recipient=True, field='production_site_certificate'))
 
     # SUPPLIER CERT
@@ -213,7 +220,7 @@ def check_certificates(prefetched_data, lot, errors):
         else:
             # certificate is set and exists. is it valid?
             c = prefetched_data['certificates'][cert]
-            if c.valid_until < lot.delivery_date:
+            if c["valid_until"] < lot.delivery_date:
                  errors.append(generic_error(error=CarbureCertificatesErrors.EXPIRED_SUPPLIER_CERT, lot=lot, display_to_recipient=True, field='supplier_certificate'))
 
     # check if mapping of SUPPLIER CERT / ENTITY has been rejected by admin
@@ -367,6 +374,9 @@ def sanity_check_mandatory_fields(lot):
     is_valid = True
     today = datetime.date.today()
     errors = []
+
+    if lot.lot_status == CarbureLot.FLUSHED:
+        return True, []
 
     if not lot.volume:
         errors.append(generic_error(error=CarbureSanityCheckErrors.MISSING_VOLUME, lot=lot, field='volume', is_blocking=True))
